pyproject.toml
[project]
name = "fastapi-views"
version = "1.5.1"
description = "FastAPI Class Views and utilities"
readme = "README.md"
requires-python = ">=3.10,<4.0"
license = "Apache-2.0"
keywords = ["fastapi", "views", "asyncio"]

authors = [{ name = "RaRhAeu", email = "rarha_eu@protonmail.com" }]
classifiers = [
"Operating System :: OS Independent",
"Programming Language :: Python :: 3",
"Programming Language :: Python",
"Topic :: Internet",
"Topic :: Software Development :: Libraries :: Application Frameworks",
"Topic :: Software Development :: Libraries :: Python Modules",
"Topic :: Software Development :: Libraries",
"Topic :: Software Development",
"Typing :: Typed",
"Development Status :: 4 - Beta",
"Environment :: Web Environment",
"Framework :: AsyncIO",
"Framework :: FastAPI",
"Framework :: Pydantic",
"Framework :: Pydantic :: 2",
"Intended Audience :: Developers",
"Programming Language :: Python :: Implementation :: CPython",
"Programming Language :: Python :: Implementation :: PyPy",
"Programming Language :: Python :: 3 :: Only",
"Programming Language :: Python :: 3.10",
"Programming Language :: Python :: 3.11",
"Programming Language :: Python :: 3.12",
"Programming Language :: Python :: 3.13",
"Programming Language :: Python :: 3.14",
"Topic :: Internet :: WWW/HTTP :: HTTP Servers",
"Topic :: Internet :: WWW/HTTP",
]
dependencies = ["fastapi", "pydantic>2.0,<3"]

[project.urls]
Documentation = "https://github.com/asynq-io/fastapi-views#readme"
Issues = "https://github.com/asynq-io/fastapi-views/issues"
Source = "https://github.com/asynq-io/fastapi-views"

[project.scripts]
fastapi-views = "fastapi_views.cli:cli"

[project.optional-dependencies]
uvloop = ["uvloop"]
prometheus = ["starlette-exporter"]
uvicorn = ["uvicorn"]
opentelemetry = ["opentelemetry-instrumentation-fastapi"]
cli = ["typer"]
all = ["uvloop", "starlette-exporter", "uvicorn", "opentelemetry-instrumentation-fastapi", "typer"]

[dependency-groups]
dev = [
    {include-group = "lint"},
    {include-group = "test"},
    {include-group = "docs"},
    "ipython",
]

test = [
    "asgi-lifespan",
    "httpx",
    "pytest",
    "pytest-cov",
    "pytest-sugar",
    "pytest-repeat",
    "pytest-timeout",
    "sqlalchemy>=2.0.48",
]

lint = [
    "mypy",
    "ruff",
    "bandit",
    "types-pyyaml>=6.0.12.20250915",
]

docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocs-autorefs",
    "mkdocstrings[python]",
    "watchdog",
]

[tool.uv.build-backend]
module-root = ""

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"


[tool.pytest.ini_options]
timeout = 3
addopts = "--cov=./fastapi_views"
testpaths = ["./tests"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.mypy]
python_version = "3.10"
# strict = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
follow_imports = "normal"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
warn_no_return = false
warn_unused_ignores = true

[tool.ruff]
target-version = "py310"

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["ALL"]

ignore = [
    "A002",
    "ANN401",
    "COM812",
    "D",
    "B008",     # do not perform function calls in argument defaults
    "E501",     # line too long
    "PLC0415",
    "N818",
    "PTH123",
    "PLW2901"
]


[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN001", "ANN002", "ANN003", "ANN201", "ANN202", "SLF001", "PLR2004", "ARG002"]
"examples/*" = ["S101", "ANN001", "ANN201", "INP001", "T201", "EM101", "TRY003"]

[tool.bandit]
skips = ['B101']
